project/requirements.txt
celery==5.0.5
fastapi==0.103.2
flower==2.0.1
uvicorn==0.23.2
yt-dlp==2023.7.6
ffmpeg-python==0.2.0
spleeter==2.4.0
pymongo==4.5.0
redis==5.0.1
whisperx @ git+https://github.com/m-bain/whisperx@main
dnspython==2.3.0